% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asie.r
\name{plot3}
\alias{plot3}
\title{Plot 3 function}
\usage{
plot3(x, y)
}
\description{
Generic function for plotting of R objects. For more details about the graphical parameter arguments, see par.
For simple scatter plots, plot.default will be used. However, there are plot methods for many R objects, including functions, data.frames, density objects, etc. Use methods(plot) and the documentation for these.
}
\examples{
plot3()
}
\keyword{plot}
