% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ranking.R
\name{rank_table}
\alias{rank_table}
\title{Create a table ranking countries based on variable of interest}
\usage{
rank_table(variable, year_select)
}
\arguments{
\item{variable}{Character string giving the variable of interest for ranking}

\item{year_select}{Year to filter to. Default is 2019}
}
\value{
a DT table
}
\description{
Compute the numeric values, \code{rank}, needed to organize table
}
